refactor: header assets and component
